hookUpdate: We have static squares
