+++ src/components/Game.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Container from "react-bootstrap/Container";
 import GameOptions from "./GameOptions";
 import GameBoard from "./GameBoard";
