+++ tsconfig.json
@@ -17,6 +17,7 @@
       }
     },
     "include": ["src/**/*.ts", "src/**/*.tsx"],
+    "files": ["custom.d.ts"],
     "exclude": ["node_modules"]
   }
